Initial file split and Match inference
